Change to class pattern
